request: more detail on CI config for prereleases
Hi! Sorry if this is super weird! I was wondering if you can elaborate on "CI runs and releases the one-off version (1.3.0-test-cool-new-feature.1)"?

I'm currently trying to recreate this in a GitHub action, in which I've 1) created a new `x.x.x-${name-of-branch}` branch, 2) push to it origin and 3) run `npx semantic-release`. Now, I'm trying to figure out how to get around semantic-release returning: `This test run was triggered on the branch <ORIGINAL_BRANCH>, while semantic-release is configured to only publish from main, beta, <GENERATED_BRANCH>, therefore a new version won’t be published.`